external/unified-cloud-services/requirements.txt
# Unified Cloud Services Dependencies
# Generated from pyproject.toml

# Core dependencies
pandas>=2.0.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.13.0
google-cloud-secret-manager>=2.16.0
google-auth>=2.25.0
pyarrow>=14.0.1
numpy>=1.21.0,<2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Byte-range streaming for Parquet files (GCS)
fsspec>=2024.2.0
gcsfs>=2024.2.0

# System monitoring dependencies (for performance_monitor.py, memory_monitor.py)
psutil>=5.9.0

# Web3 dependencies (for web3_client_pool.py)
web3>=6.0.0

# Development dependencies (optional)
# Uncomment to install dev dependencies:
# pytest>=7.4.3
# pytest-asyncio>=0.21.1
# pytest-cov>=4.1.0
# black>=23.11.0
# isort>=5.12.0
# mypy>=1.7.1
